adds functions to core legend to prepare it to be rendered
Adds the function initializeGeometry and render to the legend
model. The initializeGeometry method preforms the computations
needed for the legend to be drawn and stores them in new
attributes. The render method figures out the locations at
which each component needs to be drawn and calls functions
that individual drivers need to implement to draw the legend.