adds a method to the legend which determines if it should be drawn
adds specs

changes the 'visible' attr to allow for null, and updates its default. Adds a
method 'determineVisibility' which returns the value of the 'visible' attr if
it was set to a Boolean or returns a Boolean based on the number of plots if
the 'visible' attr was set to null, ie if it was not set in the mugl.